added coin api support

+++ backend/src/main/java/com/wild/backend/api/coins/CoinData.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -15,5 +14,5 @@ public class CoinData {
     private String id;
     private String name;
     private String symbol;
-    private BigDecimal price;
+    private Double price;
 }
